update: split tests
